`LlmSchemaFactory.fixDatabasePlugin()`
