+++ packages/agent/src/orchestrate/interface/orchestrateInterfaceSchemaReview.ts
@@ -18,6 +18,7 @@ import { IAutoBeInterfaceSchemaReviewApplication } from "./structures/IAutoBeInt
 import { JsonSchemaFactory } from "./utils/JsonSchemaFactory";
 import { JsonSchemaNamingConvention } from "./utils/JsonSchemaNamingConvention";
 import { JsonSchemaValidator } from "./utils/JsonSchemaValidator";
+import { LlmSchemaFactory } from "./utils/LlmSchemaFactory";
 import { fulfillJsonSchemaErrorMessages } from "./utils/fulfillJsonSchemaErrorMessages";
 
 interface IConfig {
@@ -252,7 +253,7 @@ function createController(
         ] as ILlmSchema.IObject
       ).properties.content as ILlmSchema.IReference
     ).$ref = "AutoBeOpenApi.IJsonSchemaDescriptive.IObject";
-  JsonSchemaFactory.fixPlugin(
+  LlmSchemaFactory.fixDatabasePlugin(
     ctx.state(),
     application.functions[0].parameters.$defs,
   );
